Trinity/requirements.txt
aiogram==3.7.0
apscheduler==3.10.4
python-dotenv>=1.0.1
